Add C globals setup code
For now I'm just hardcoding a single session, BLE connection, and one
GATT Service in the C side, that I'll re-use on the Go side. This is not
ideal because it means you can only have a single
session/connection/service going on at the same time, but for now it's
good enough.